backend/requirements.txt
anthropic>=0.40.0
supabase>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
